requirements.txt
certifi==2020.11.8
chardet==3.0.4
dandelion-eu==0.3.2
idna==2.10
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
SpeechRecognition==3.8.1
urllib3==1.26.2
